+++ parquet-parqour/src/main/java/org/apache/parquet/parqour/ingest/ffreader/dictionary/RLEFixedBinaryDictionaryFastForwardReader.java
@@ -34,14 +34,14 @@ public final class RLEFixedBinaryDictionaryFastForwardReader extends FastForward
     super(info, values);
 
     this.dictionaryEntries = this.readDictionaryEntries(info);
-    this.dictionaryEntriesAsStrings = new String[(int) info.dictionaryPage().entryCount()];
+    this.dictionaryEntriesAsStrings = new String[info.dictionaryPage().entryCount()];
     this.segment = PackedEncodingSegmentReader.createPackedEncodingSegmentReader(data, info.contentOffset(), expandToBitWidth(info));
   }
 
-
+  // Todo: centralize this.
   private byte[][] readDictionaryEntries(DataPageInfo info) {
     DictionaryPageInfo dictionaryPage = info.dictionaryPage();
-    int dictionarySize = (int) dictionaryPage.entryCount();
+    int dictionarySize = dictionaryPage.entryCount();
     int typeLength = info.typeLength();
     byte[] dictionaryData = dictionaryPage.data();
